Fixed major bug in interpolation of the potential perturbation derivatives (dpsi_dx and dpsi_dy). Also, minor additions/modifications here and there.

+++ analysis_tools/finalize.sh
@@ -51,7 +51,9 @@ php get_par_table.php $path $run $step
 
 
 mv all.png $target
-mv corner.pdf $target
+if [ -f corner.pdf ]; then
+    mv corner.pdf $target
+fi
 mv report.pdf $target
 mv table_pars.txt $target/
